# docs/openapi.yaml
...
openapi: 3.0.1
servers:
  - url: 'http://localhost:8081'
    description: 'Generated server url'
paths:
...
  /register:
    post:
      tags:
        - Регистрация
      summary: 'Регистрация пользователя'
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Register'
        required: true
      responses:
        '400':
          description: 'Bad Request'
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/JwtAuthenticationResponse'
  /login:
    post:
      tags:
        - Аутентификация
      summary: 'Аутентификация пользователя'
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/JwtAuthenticationResponse'
        '401':
          description: Unauthorized
...
info:
  title: Bank Card Management System
  version: 0.0.1
components:
  schemas:
...
    Register:
      required:
        - password
        - role
        - username
      type: object
      properties:
        username:
          maxLength: 32
          minLength: 4
          type: string
          description: логин
        password:
          maxLength: 16
          minLength: 8
          type: string
          description: пароль
        role:
          type: string
          description: 'роль пользователя'
          enum:
            - ADMIN
            - USER
      description: 'Форма регистрации'
    JwtAuthenticationResponse:
      type: object
      properties:
        token:
          type: string
          description: 'Токен доступа'
          example: 'eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJhZG1pbiIsImV4cCI6MTYyMjUwNj...'
      description: 'Ответ c токеном доступа'
    Login:
      required:
        - password
        - username
      type: object
      properties:
        username:
          maxLength: 32
          minLength: 4
          type: string
          description: логин
        password:
          maxLength: 16
          minLength: 8
          type: string
          description: пароль
      description: 'Форма аутентификации'